# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MLM Direct Selling Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ReferralInfo {
  referrerId: string;
  maskedName: string;
  isValid: boolean;
}

export type RegistrationInputPackageType = typeof RegistrationInputPackageType[keyof typeof RegistrationInputPackageType];


export const RegistrationInputPackageType = {
  starter: 'starter',
  pro: 'pro',
  elite: 'elite',
} as const;

export interface RegistrationInput {
  email: string;
  password: string;
  confirmPassword: string;
  referrerId: string;
  firstName: string;
  lastName: string;
  mobileNumber: string;
  address: string;
  countryCode: string;
  packageType: RegistrationInputPackageType;
  agreedToTerms?: boolean;
}

export interface PendingRegistration {
  id: number;
  email: string;
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  pending: 'pending',
  active: 'active',
  suspended: 'suspended',
} as const;

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  distributor: 'distributor',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  mobileNumber?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  countryCode?: string | null;
  status: UserStatus;
  role: UserRole;
  referralCode: string;
  /** @nullable */
  referrerId?: string | null;
  /** @nullable */
  packageType?: string | null;
  leftBv?: number;
  rightBv?: number;
  createdAt: string;
}

export interface UserUpdate {
  firstName?: string;
  lastName?: string;
  mobileNumber?: string;
  address?: string;
  countryCode?: string;
}

export type UserStatusUpdateStatus = typeof UserStatusUpdateStatus[keyof typeof UserStatusUpdateStatus];


export const UserStatusUpdateStatus = {
  active: 'active',
  suspended: 'suspended',
} as const;

export interface UserStatusUpdate {
  status: UserStatusUpdateStatus;
}

/**
 * @nullable
 */
export type TreeNodeLeg = typeof TreeNodeLeg[keyof typeof TreeNodeLeg] | null;


export const TreeNodeLeg = {
  left: 'left',
  right: 'right',
} as const;

export interface TreeNode {
  id: number;
  userId: number;
  firstName: string;
  lastName: string;
  status: string;
  /** @nullable */
  packageType?: string | null;
  /** @nullable */
  sponsorId?: number | null;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  leg?: TreeNodeLeg;
  /** @nullable */
  leftChildId?: number | null;
  /** @nullable */
  rightChildId?: number | null;
  leftBv: number;
  rightBv: number;
  depth: number;
}

export type NodePlacementInputLeg = typeof NodePlacementInputLeg[keyof typeof NodePlacementInputLeg];


export const NodePlacementInputLeg = {
  left: 'left',
  right: 'right',
} as const;

export interface NodePlacementInput {
  userId: number;
  newParentId: number;
  leg: NodePlacementInputLeg;
}

export interface NetworkStats {
  totalNodes: number;
  activeNodes: number;
  pendingNodes: number;
  totalLeftBv: number;
  totalRightBv: number;
}

export type CommissionType = typeof CommissionType[keyof typeof CommissionType];


export const CommissionType = {
  direct_referral: 'direct_referral',
  binary_match: 'binary_match',
} as const;

export interface Commission {
  id: number;
  userId: number;
  type: CommissionType;
  amount: number;
  sourceUserId: number;
  /** @nullable */
  sourceUserName?: string | null;
  /** @nullable */
  bvMatched?: number | null;
  createdAt: string;
}

export interface CommissionStats {
  totalEarned: number;
  directReferralTotal: number;
  binaryMatchTotal: number;
  pendingCycles: number;
  leftBv: number;
  rightBv: number;
  matchedBv: number;
  residualLeftBv: number;
  residualRightBv: number;
}

export interface Wallet {
  id: number;
  userId: number;
  totalEarned: number;
  totalSpent: number;
  availableBalance: number;
  hasPin: boolean;
  createdAt: string;
}

export type WalletTransactionType = typeof WalletTransactionType[keyof typeof WalletTransactionType];


export const WalletTransactionType = {
  credit: 'credit',
  debit: 'debit',
} as const;

export interface WalletTransaction {
  id: number;
  walletId: number;
  type: WalletTransactionType;
  amount: number;
  description: string;
  /** @nullable */
  referenceId?: string | null;
  createdAt: string;
}

export interface WalletPinInput {
  pin: string;
}

export interface WalletPinUpdate {
  pin: string;
  /** @nullable */
  currentPin?: string | null;
}

export interface PinVerifyResult {
  valid: boolean;
}

export type CouponStatus = typeof CouponStatus[keyof typeof CouponStatus];


export const CouponStatus = {
  active: 'active',
  redeemed: 'redeemed',
  expired: 'expired',
} as const;

export interface Coupon {
  id: number;
  userId: number;
  code: string;
  amount: number;
  status: CouponStatus;
  /** @nullable */
  redeemedAt?: string | null;
  createdAt: string;
}

export interface CouponInput {
  amount: number;
  pin: string;
}

export interface CouponRedeemInput {
  code: string;
}

export interface FinancialStats {
  totalInflow: number;
  totalCommissionPaid: number;
  totalExpenses: number;
  totalOutflow: number;
  washRatio: number;
  isInsolvent: boolean;
  retentionPool: number;
  overheadPool: number;
  commissionPool: number;
}

export interface WashResetInput {
  adminPin: string;
  /** @nullable */
  reason?: string | null;
}

export interface WashResetResult {
  success: boolean;
  walletsReset: number;
  commissionsReset: number;
  totalAmountWiped: number;
  executedAt: string;
}

export interface WashEvent {
  id: number;
  executedBy: number;
  walletsReset: number;
  totalAmountWiped: number;
  /** @nullable */
  reason: string | null;
  createdAt: string;
}

export interface GlobalBvStats {
  globalLeftBv: number;
  globalRightBv: number;
  totalMatchedCycles: number;
  totalBonusPaid: number;
  activeMemberCount: number;
}

export interface DashboardSummary {
  userId: number;
  leftBv: number;
  rightBv: number;
  directReferrals: number;
  totalDownline: number;
  walletBalance: number;
  totalEarned: number;
  pendingCycles: number;
  nextCycleLeftNeeded: number;
  nextCycleRightNeeded: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  registration: 'registration',
  commission_earned: 'commission_earned',
  binary_match: 'binary_match',
  coupon_created: 'coupon_created',
  coupon_redeemed: 'coupon_redeemed',
  wash_reset: 'wash_reset',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  /** @nullable */
  amount: number | null;
  /** @nullable */
  relatedUserId?: number | null;
  /** @nullable */
  relatedUserName?: string | null;
  createdAt: string;
}

export type VerifyReferralParams = {
refId: string;
};

export type ListUsersParams = {
status?: ListUsersStatus;
search?: string;
};

export type ListUsersStatus = typeof ListUsersStatus[keyof typeof ListUsersStatus];


export const ListUsersStatus = {
  pending: 'pending',
  active: 'active',
  suspended: 'suspended',
} as const;

export type GetNetworkTreeParams = {
/**
 * @nullable
 */
rootUserId?: number | null;
};

export type ListCommissionsParams = {
/**
 * @nullable
 */
userId?: number | null;
type?: ListCommissionsType;
};

export type ListCommissionsType = typeof ListCommissionsType[keyof typeof ListCommissionsType];


export const ListCommissionsType = {
  direct_referral: 'direct_referral',
  binary_match: 'binary_match',
} as const;

export type GetCommissionStatsParams = {
/**
 * @nullable
 */
userId?: number | null;
};