# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MLM Direct Selling Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/verify-referral:
    get:
      operationId: verifyReferral
      tags: [auth]
      summary: Validate referral ID and return masked sponsor name
      parameters:
        - name: refId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Valid referral
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralInfo"
        "404":
          description: Not found
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user (creates pending registration)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Pending registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingRegistration"
        "400":
          description: Invalid input
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ── USERS ─────────────────────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, suspended]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      summary: Suspend or activate a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/approve-payment:
    post:
      operationId: approvePayment
      tags: [users]
      summary: Manually approve payment and activate user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── NETWORK ───────────────────────────────────────────────────────────────
...
  /admin/wash-reset:
    post:
      operationId: executeWashReset
      tags: [admin]
      summary: Execute the Wash Reset — wipes unredeemed wallet balances
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WashResetInput"
      responses:
        "200":
          description: Wash reset executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WashResetResult"
...
components:
  schemas:
...
    ReferralInfo:
      type: object
      required: [referrerId, maskedName, isValid]
      properties:
        referrerId:
          type: string
        maskedName:
          type: string
        isValid:
          type: boolean
    RegistrationInput:
      type: object
      required: [email, password, confirmPassword, referrerId, firstName, lastName, mobileNumber, address, countryCode, packageType]
      properties:
        email:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
        referrerId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        mobileNumber:
          type: string
        address:
          type: string
        countryCode:
          type: string
        packageType:
          type: string
          enum: [starter, pro, elite]
        agreedToTerms:
          type: boolean
    PendingRegistration:
      type: object
      required: [id, email, status]
      properties:
        id:
          type: integer
        email:
          type: string
        status:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    # ── USER SCHEMAS ──────────────────────────────────────────────────────
    User:
      type: object
      required: [id, email, firstName, lastName, status, role, referralCode, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        mobileNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        countryCode:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, suspended]
        role:
          type: string
          enum: [distributor, admin]
        referralCode:
          type: string
        referrerId:
          type: ["string", "null"]
        packageType:
          type: ["string", "null"]
        leftBv:
          type: number
        rightBv:
          type: number
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        mobileNumber:
          type: string
        address:
          type: string
        countryCode:
          type: string
    UserStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, suspended]

    # ── NETWORK SCHEMAS ───────────────────────────────────────────────────
...
    WashResetInput:
      type: object
      required: [adminPin]
      properties:
        adminPin:
          type: string
        reason:
          type: ["string", "null"]
    WashResetResult:
      type: object
      required: [success, walletsReset, commissionsReset, totalAmountWiped, executedAt]
      properties:
        success:
          type: boolean
        walletsReset:
          type: integer
        commissionsReset:
          type: integer
        totalAmountWiped:
          type: number
        executedAt:
          type: string
...