# lib/api-zod/src/generated/types/referralInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MLM Direct Selling Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ReferralInfo {
  referrerId: string;
  maskedName: string;
  isValid: boolean;
}