# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MLM Direct Selling Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationInputPackageType } from './registrationInputPackageType';

export interface RegistrationInput {
  email: string;
  password: string;
  confirmPassword: string;
  referrerId: string;
  firstName: string;
  lastName: string;
  mobileNumber: string;
  address: string;
  countryCode: string;
  packageType: RegistrationInputPackageType;
  agreedToTerms?: boolean;
}