# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MLM Direct Selling Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  mobileNumber?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  countryCode?: string | null;
  status: UserStatus;
  role: UserRole;
  referralCode: string;
  /** @nullable */
  referrerId?: string | null;
  /** @nullable */
  packageType?: string | null;
  leftBv?: number;
  rightBv?: number;
  createdAt: string;
}