# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MLM Direct Selling Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  firstName?: string;
  lastName?: string;
  mobileNumber?: string;
  address?: string;
  countryCode?: string;
}